third_party/blink/web_tests/platform/generic/editing/pasteboard/drag-drop-list-expected.txt

This tests dropping content onto a list items content when the content is wrapped in a span.  You should see 'DropAboveMe DragMe'. See <rdar://4861080>, but note that the expected behavior is different than what's described there, since the rendering bug <rdar://4907469> has been fixed.
| "
"
| <li>
|   id="li"
|   style="border: 1px solid black;"
|   <span>
|     <div>
|       id="contents"
|       style="border: 1px solid red;"
|       "<#selection-anchor>DragMe <#selection-focus>DropAboveMe"
| "
"
| "
"
